[Refactor] 리스트로 받기 변경

+++ src/main/java/kr/inuappcenterportal/inuportal/service/FolderService.java
@@ -12,8 +12,6 @@ import kr.inuappcenterportal.inuportal.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.webjars.NotFoundException;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -49,32 +47,41 @@ public class FolderService {
 
     @Transactional
     public Long insertPost(FolderPostDto folderPostDto){
-        Folder folder = folderRepository.findById(folderPostDto.getFolderId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_NOT_FOUND));
-        Post post = postRepository.findById(folderPostDto.getPostId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.POST_NOT_FOUND));
-        Member member =memberRepository.findById(folder.getMember().getId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.USER_NOT_FOUND));
-        Scrap scrap = scrapRepository.findByMemberAndPost(member,post).orElseThrow(()->new MyNotFoundException(MyErrorCode.SCRAP_NOT_FOUND));
-        if(folderPostRepository.existsByFolderAndPost(folder,post)){
-            System.out.println(folder.getName());
-            System.out.println(post.getId());
-            throw new MyDuplicateException(MyErrorCode.POST_DUPLICATE_FOLDER);
+        if(folderPostDto.getPostId().size()==0){
+            throw new MyNotFoundException(MyErrorCode.POST_SCRAP_LIST_NOT_FOUND);
         }
-        return folderPostRepository.save(FolderPost.builder().post(post).folder(folder).scrap(scrap).build()).getId();
+        Folder folder = folderRepository.findById(folderPostDto.getFolderId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_NOT_FOUND));
+        Member member =memberRepository.findById(folder.getMember().getId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.USER_NOT_FOUND));
+        for(Long id:folderPostDto.getPostId()){
+            Post post = postRepository.findById(id).orElseThrow(()->new MyNotFoundException(MyErrorCode.POST_NOT_FOUND));
+            Scrap scrap = scrapRepository.findByMemberAndPost(member,post).orElseThrow(()->new MyNotFoundException(MyErrorCode.SCRAP_NOT_FOUND));
+            if(folderPostRepository.existsByFolderAndPost(folder,post)){
+                throw new MyDuplicateException(MyErrorCode.POST_DUPLICATE_FOLDER);
+            }
+            folderPostRepository.save(FolderPost.builder().post(post).folder(folder).scrap(scrap).build());
+        }
+        return folderPostDto.getFolderId();
     }
 
     @Transactional
     public void deleteInFolder(FolderPostDto folderPostDto){
+        if(folderPostDto.getPostId().size()==0){
+            throw new MyNotFoundException(MyErrorCode.POST_SCRAP_LIST_NOT_FOUND);
+        }
         Folder folder = folderRepository.findById(folderPostDto.getFolderId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_NOT_FOUND));
-        Post post = postRepository.findById(folderPostDto.getFolderId()).orElseThrow(()->new MyNotFoundException(MyErrorCode.POST_NOT_FOUND));
-        folderPostRepository.delete(folderPostRepository.findByFolderAndPost(folder,post).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_OR_POST_NOT_FOUND)));
+        for(Long id:folderPostDto.getPostId()){
+            Post post = postRepository.findById(id).orElseThrow(()->new MyNotFoundException(MyErrorCode.POST_NOT_FOUND));
+            folderPostRepository.delete(folderPostRepository.findByFolderAndPost(folder,post).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_OR_POST_NOT_FOUND)));
+        }
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<FolderResponseDto> getFolder(Long memberId){
         Member member = memberRepository.findById(memberId).orElseThrow(()->new MyNotFoundException(MyErrorCode.USER_NOT_FOUND));
         return folderRepository.findAllByMember(member).stream().map(FolderResponseDto::new).collect(Collectors.toList());
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<PostListResponseDto> getPostInFolder(Long folderId){
         Folder folder = folderRepository.findById(folderId).orElseThrow(()->new MyNotFoundException(MyErrorCode.FOLDER_NOT_FOUND));
         return folderPostRepository.findAllByFolder(folder).stream().map(file->postService.getPostListResponseDto(file.getPost())).collect(Collectors.toList());
